extend observer, remove old solver and environment
