+++ .gitignore
@@ -7,6 +7,8 @@ diff_patch_search-0.0.1*
 repo-description.txt
 playground*
 temp/
+*.pickle
+old_*
 
 # Created by https://www.toptal.com/developers/gitignore/api/visualstudiocode,python,jupyternotebooks
 # Edit at https://www.toptal.com/developers/gitignore?templates=visualstudiocode,python,jupyternotebooks
