+++ pyproject.toml
@@ -32,6 +32,8 @@ typing_extensions = "^4.11.0"
 instructor = "1.2.5"
 anthropic = "0.25.8"
 tenacity = "8.3.0"
+tree_sitter_languages = "^1.10.2"
+tree-sitter = "^0.21.3"
 
 
 [build-system]
