+++ se_gym/__init__.py
@@ -1,7 +1,10 @@
-from .solver import Solver
 from .call_api import call_model
-from .environment import Environment
-from .observations import *
+from .observations import (
+    ManualObserver,
+    VectorStoreObserver,
+    TreeSitterObeserver,
+    CodeMapObserver,
+)
 from .caller import Sampler
 from .openai_lmu import get_openai_client
 from .executor import apply_patch, apply_patch_and_test, MalformedPatchException
